source_code/requirements.txt
fastapi
uvicorn
kafka-python
redis
mongoengine
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-kafka-python
opentelemetry-instrumentation-pymongo
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
